Add kubernetes manifest files
This will add a directory in the scripts dir which contains all of
the necessary kubernetes manifest files to deploy onto a user's
managed kubernetes platform of choice. These have been written to
be platform-agnostic, but have not yet been tested for use with
Minikube.

Documentation will need to be added, but I want to get opinions
and have discussions with the geonode community before moving
further. I have a running deployment using these configs here:

https://eggshell.us-south.containers.appdomain.cloud/

Signed-off-by: Cullen Taylor <mctaylor@us.ibm.com>